This modification code defines a function to handle click events on a set of icon filters.
The object called turnOn is for maps each filter's id to a function that will turn it on.
An isActive function is defined to handle the click event on any of the icon filters. This function uses the turnOn object to call the corresponding function to turn on the clicked filter. Additionally, it calls a disable function if the low-cost filter is already active to clear it.