Function to display 3 img in the popular hotels
This is a JavaScript function that takes in two parameters: imgs and popularContainer. The imgs parameter is an array of objects that contain information about popular hotels, such as their name, value, and URL of their image. The popularContainer parameter is the HTML container where these images will be displayed.

The function loops through the imgs array using the forEach() method and creates an HTML element for each image. It uses template literals to create a string that includes the image URL, name, value, and a star rating. The if statement is used to limit the number of images displayed to the value of the limit constant, which is set to 3 in this example.

Finally, the function appends the created HTML element to the popularContainer using the innerHTML property. This means that the previous content of the popularContainer will be replaced with the new HTML generated by the function.